improved first version of third & fourth component
their was a mistake in the previous commit, i said it was second and third component but it was actually the third and fourth component which are the task display and add tasks frame respectively. The task add frame now has a help description in order to guide the user through adding tasks. Another improvement is I fixed all the placements so the program will look good. With a barely functioning logout system.